수정 SW Expert Academy - 백만 장자 프로젝트
https://swexpertacademy.com/main/code/problem/problemDetail.do?contestProbId=AV5LrsUaDxcDFAXc&categoryId=AV5LrsUaDxcDFAXc&categoryType=CODE

int to long long type

+++ Study-Algorithm/백만 장자 프로젝트.cpp
@@ -5,12 +5,12 @@
 
 using namespace std;
 
-int solution(vector<int>::iterator begin, vector<int>::iterator end) {
+long long solution(vector<int>::iterator begin, vector<int>::iterator end) {
 	if (begin == end) {
 		return 0;
 	}
 	vector<int>::iterator maxIt = max_element(begin, end);
-	int gain = 0;
+	long long gain = 0;
 	while (begin != maxIt) {
 		gain += *maxIt - *begin++;
 	}
@@ -21,7 +21,7 @@ int main(void) {
 	int t;
 	cin >> t;
 	for (int test = 0; test < t; test++) {
-		int answer = 0;
+		long long answer = 0;
 		int days;
 		vector<int> prices;
 		cin >> days;
